Refactor useConfigManager hook
Refactor `src/hooks/useConfigManager.ts` into smaller, more focused files to improve maintainability.
