+++ src/hooks/config/useConfigValidation.ts
@@ -1,0 +1,51 @@
+
+import { useState } from 'react';
+import { RF4SYamlConfig } from '../../types/config';
+
+export const useConfigValidation = () => {
+  const [isTestMode, setIsTestMode] = useState(false);
+  const [testResults, setTestResults] = useState<any>(null);
+
+  const testConfiguration = async (config: RF4SYamlConfig) => {
+    setIsTestMode(true);
+    console.log('Testing configuration...');
+    
+    // Simulate configuration testing
+    await new Promise(resolve => setTimeout(resolve, 2000));
+    
+    const results = {
+      valid: true,
+      warnings: [],
+      errors: [],
+      suggestions: [
+        'Consider increasing SPOOL_CONFIDENCE to 0.98 for better accuracy',
+        'CAST_DELAY could be optimized for faster fishing cycles'
+      ]
+    };
+    
+    setTestResults(results);
+    setIsTestMode(false);
+    return results;
+  };
+
+  const aiOptimizeConfig = async (config: RF4SYamlConfig) => {
+    console.log('AI optimizing configuration...');
+    
+    // Simulate AI optimization
+    await new Promise(resolve => setTimeout(resolve, 3000));
+    
+    const optimizedConfig = { ...config };
+    optimizedConfig.SCRIPT.SPOOL_CONFIDENCE = 0.98;
+    optimizedConfig.SCRIPT.RANDOM_CAST_PROBABILITY = 0.15;
+    optimizedConfig.FRICTION_BRAKE.INITIAL = 25;
+    
+    return optimizedConfig;
+  };
+
+  return {
+    isTestMode,
+    testResults,
+    testConfiguration,
+    aiOptimizeConfig
+  };
+};
